feat: 프로젝트 추가 모든 페이지에서 되게 설정
