+++ src/layouts/App/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import '@layouts/App/App.css';
 import loadable from '@loadable/component';
 import ReleaseNoteDetail from '@components/ReleaseNote/ReleaseNoteDetail';
+import AddProject from '@components/AddProject';
 
 const InitialPage = loadable(() => import('@pages/InitialPage'));
 const Main = loadable(() => import('@layouts/Main'));
@@ -21,33 +22,36 @@ const ReleaseNoteEdit = loadable(() => import('@components/ReleaseNote/ReleaseNo
 
 function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Navigate replace to="/initial" />} />
-        <Route path="/main" element={<Main />} />
-        <Route path="/initial" element={<InitialPage />} />
-        <Route path={'/project/:projectId/manage'} element={<ManagePage />} />
-        <Route path="/mypage" element={<MyPage />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path={'/project/:projectId/issue/*'}>
-          <Route path="" element={<IssuePage />} />
-          <Route path=":issueId" element={<IssueDetail />} />
-          <Route path=":issueId/editIssue" element={<EditIssue />} />
-        </Route>
-        <Route path="/issue/createissue" element={<CreateIssue />} />
-        <Route path={'/project/:projectId/projectinfo'} element={<ProjectInfo />} />
-        <Route path={'/project/:projectId/releasenote/*'}>
-          <Route path="" index element={<ReleaseNotePage />} />
-          <Route path=":releaseId" element={<ReleaseNoteDetail />} />
-          <Route path=":releaseId/edit" element={<ReleaseNoteEdit />} />
-        </Route>
-        <Route path={'/project/:projectId/workspace/*'}>
-          <Route path="" element={<WorkSpacePage />} />
-          <Route path=":workspaceId" element={<WorkSpaceDetail />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Navigate replace to="/initial" />} />
+          <Route path="/main" element={<Main />} />
+          <Route path="/initial" element={<InitialPage />} />
+          <Route path={'/project/:projectId/manage'} element={<ManagePage />} />
+          <Route path="/mypage" element={<MyPage />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/signup" element={<SignUp />} />
+          <Route path={'/project/:projectId/issue/*'}>
+            <Route path="" element={<IssuePage />} />
+            <Route path=":issueId" element={<IssueDetail />} />
+            <Route path=":issueId/editIssue" element={<EditIssue />} />
+            <Route path=":issueId/createissue" element={<CreateIssue />} />
+          </Route>
+          <Route path={'/project/:projectId/projectinfo'} element={<ProjectInfo />} />
+          <Route path={'/project/:projectId/releasenote/*'}>
+            <Route path="" index element={<ReleaseNotePage />} />
+            <Route path=":releaseId" element={<ReleaseNoteDetail />} />
+            <Route path=":releaseId/edit" element={<ReleaseNoteEdit />} />
+          </Route>
+          <Route path={'/project/:projectId/workspace/*'}>
+            <Route path="" element={<WorkSpacePage />} />
+            <Route path=":workspaceId" element={<WorkSpaceDetail />} />
+          </Route>
+        </Routes>
+      </BrowserRouter>
+      <AddProject />
+    </>
   );
 }
